# Increase Appender buffer size, and improve appender performance when appending strings