--- a/src/common/operator/string_cast.cpp
+++ b/src/common/operator/string_cast.cpp
@@ -146,4 +146,9 @@ duckdb::string_t StringCast::Operation(timestamp_t input, Vector &vector) {
 	return result;
 }
 
+template <>
+duckdb::string_t StringCast::Operation(duckdb::string_t input, Vector &result) {
+	return StringVector::AddStringOrBlob(result, input);
+}
+
 } // namespace duckdb