--- a/src/include/duckdb/common/types/chunk_collection.hpp
+++ b/src/include/duckdb/common/types/chunk_collection.hpp
@@ -48,6 +48,9 @@ public:
 	//! Append a new DataChunk directly to this ChunkCollection
 	DUCKDB_API void Append(DataChunk &new_chunk);
 
+	//! Append a new DataChunk directly to this ChunkCollection
+	DUCKDB_API void Append(unique_ptr<DataChunk> new_chunk);
+
 	//! Append another ChunkCollection directly to this ChunkCollection
 	DUCKDB_API void Append(ChunkCollection &other);