--- a/src/include/duckdb/main/appender.hpp
+++ b/src/include/duckdb/main/appender.hpp
@@ -130,6 +130,8 @@ void DUCKDB_API Appender::Append(timestamp_t value);
 template <>
 void DUCKDB_API Appender::Append(const char *value);
 template <>
+void DUCKDB_API Appender::Append(string_t value);
+template <>
 void DUCKDB_API Appender::Append(Value value);
 template <>
 void DUCKDB_API Appender::Append(std::nullptr_t value);