--- a/src/main/appender.cpp
+++ b/src/main/appender.cpp
@@ -17,13 +17,11 @@ Appender::Appender(Connection &con, const string &schema_name, const string &tab
 	if (!description) {
 		// table could not be found
 		throw CatalogException(StringUtil::Format("Table \"%s.%s\" could not be found", schema_name, table_name));
-	} else {
-		vector<LogicalType> types;
-		for (auto &column : description->columns) {
-			types.push_back(column.type);
-		}
-		chunk.Initialize(types);
 	}
+	for (auto &column : description->columns) {
+		types.push_back(column.type);
+	}
+	InitializeChunk();
 }
 
 Appender::Appender(Connection &con, const string &table_name) : Appender(con, DEFAULT_SCHEMA, table_name) {
@@ -38,32 +36,37 @@ Appender::~Appender() {
 	}
 }
 
+void Appender::InitializeChunk() {
+	chunk = make_unique<DataChunk>();
+	chunk->Initialize(types);
+}
+
 void Appender::BeginRow() {
 }
 
 void Appender::EndRow() {
 	// check that all rows have been appended to
-	if (column != chunk.ColumnCount()) {
+	if (column != chunk->ColumnCount()) {
 		throw InvalidInputException("Call to EndRow before all rows have been appended to!");
 	}
 	column = 0;
-	chunk.SetCardinality(chunk.size() + 1);
-	if (chunk.size() >= STANDARD_VECTOR_SIZE) {
-		Flush();
+	chunk->SetCardinality(chunk->size() + 1);
+	if (chunk->size() >= STANDARD_VECTOR_SIZE) {
+		FlushChunk();
 	}
 }
 
 template <class SRC, class DST>
 void Appender::AppendValueInternal(Vector &col, SRC input) {
-	FlatVector::GetData<DST>(col)[chunk.size()] = Cast::Operation<SRC, DST>(input);
+	FlatVector::GetData<DST>(col)[chunk->size()] = Cast::Operation<SRC, DST>(input);
 }
 
 template <class T>
 void Appender::AppendValueInternal(T input) {
-	if (column >= chunk.ColumnCount()) {
+	if (column >= types.size()) {
 		throw InvalidInputException("Too many appends for chunk!");
 	}
-	auto &col = chunk.data[column];
+	auto &col = chunk->data[column];
 	switch (col.GetType().InternalType()) {
 	case PhysicalType::BOOL:
 		AppendValueInternal<T, bool>(col, input);
@@ -192,7 +195,7 @@ void Appender::Append(timestamp_t value) {
 
 template <>
 void Appender::Append(Value value) { // NOLINT: template shtuff
-	if (column >= chunk.ColumnCount()) {
+	if (column >= chunk->ColumnCount()) {
 		throw InvalidInputException("Too many appends for chunk!");
 	}
 	AppendValue(value);
@@ -200,16 +203,27 @@ void Appender::Append(Value value) { // NOLINT: template shtuff
 
 template <>
 void Appender::Append(std::nullptr_t value) {
-	if (column >= chunk.ColumnCount()) {
+	if (column >= chunk->ColumnCount()) {
 		throw InvalidInputException("Too many appends for chunk!");
 	}
-	auto &col = chunk.data[column++];
-	FlatVector::SetNull(col, chunk.size(), true);
+	auto &col = chunk->data[column++];
+	FlatVector::SetNull(col, chunk->size(), true);
 }
 
 void Appender::AppendValue(const Value &value) {
-	chunk.SetValue(column, chunk.size(), value);
+	chunk->SetValue(column, chunk->size(), value);
 	column++;
+}
+
+void Appender::FlushChunk() {
+	if (chunk->size() == 0) {
+		return;
+	}
+	collection.Append(move(chunk));
+	InitializeChunk();
+	if (collection.ChunkCount() >= FLUSH_COUNT) {
+		Flush();
+	}
 }
 
 void Appender::Flush() {
@@ -218,17 +232,18 @@ void Appender::Flush() {
 		throw InvalidInputException("Failed to Flush appender: incomplete append to row!");
 	}
 
-	if (chunk.size() == 0) {
+	FlushChunk();
+	if (collection.Count() == 0) {
 		return;
 	}
-	context->Append(*description, chunk);
+	context->Append(*description, collection);
 
-	chunk.Reset();
+	collection.Reset();
 	column = 0;
 }
 
 void Appender::Close() {
-	if (column == 0 || column == chunk.ColumnCount()) {
+	if (column == 0 || column == types.size()) {
 		Flush();
 	}
 }