This PR fixes two performance problems with the appender:

* The buffer size is increased (from 1024 entries to 102400 entries). This especially matters when the appender is used in auto-commit mode (i.e. not wrapped in a `BEGIN/COMMIT` block).
* Improve performance of appending strings, particularly from the C API, by avoiding constructing Value and string objects constantly

We also add a benchmark based on [this blog post](https://avi.im/blag/2021/fast-sqlite-inserts/), but only with 10M rows instead of 100M to avoid requiring a machine with too much memory to run the benchmarks. Results:

|       Configuration       | Time  |
|---------------------------|-------|
| In-Memory, No Primary Key | 0.97s |
| In-Memory, Primary Key    | 6.5s  |
| On-Disk, No Primary Key   | 1.66s |
| On-Disk, Primary Key      | 7.35s |

Running this locally with 100M rows confirms linear performance degradation with inserted data size:

|       Configuration       | Time  |
|---------------------------|-------|
| In-Memory, No Primary Key | 9.5s |
| In-Memory, Primary Key    | 72.9s  |
| On-Disk, No Primary Key   | 18.2s |
| On-Disk, Primary Key      | 80.4s |