@wangfenjin 2021-08-02
On my machine bench Appender10MRowsPrimaryKey takes about 12s, rust takes about 19s.

The C API shouldn't have extra work right? Then it might related to my library...