@wangfenjin 2021-08-02
Any finding is after run Appender10MRowsDiskPrimaryKey, the duckdb_benchmark_db.db file size is 203M; but after run rust test the file size is 184M. 

Maybe that's related to the random generator? But if file size is smaller benchmark should be faster, haha...