@Mytherin 2021-08-02
> On my machine bench Appender10MRowsPrimaryKey takes about 12s, rust takes about 19s.
> 
> The C API shouldn't have extra work right? Then it might related to my library...

There is a NULL check in the C API, you might want to remove that:

```cpp
template <class T>
duckdb_state duckdb_append_internal(duckdb_appender appender, T value) {
	if (!appender) {
		return DuckDBError;
	}
	auto *appender_instance = (Appender *)appender;
	try {
		appender_instance->Append<T>(value);
	} catch (...) {
		return DuckDBError;
	}
	return DuckDBSuccess;
}
```